pyproject.toml
[project]
name = "escrow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "cookiecutter>=2.6.0",
    "django-cors-headers>=4.9.0",
    "django-crispy-forms>=2.4",
    "django-tailwind[cookiecutter,honcho,reload]>=4.2.0",
    "django-vite>=3.1.0",
    "honcho>=2.0.0",
    "paystackapi>=2.1.3",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "redis>=7.0.0",
]
